Don't mutate passed in list of names
Previously, the behavior of robin.js would mutate the original list of names passed into the function. 

Passing in ['a', 'b', 'c'], the original variable would be modified to end up ['a', 'b', 'c', -1, 'b'] or something like that. 

This change will prevent the original list of names from changing.